refactor(snapshot): move Android helper presentation
