+++ src/snapshot/snapshot-presentation/android/geometry.ts
