+++ src/snapshot/snapshot-presentation/android/labels.ts
