+++ src/snapshot/snapshot-presentation/android/predicates.ts
