+++ src/snapshot/snapshot-presentation/android/structural-noise.ts
